some extension methods to try to clean up name/metadata swapping for node types
